+++ packages/terra-card/src/Card.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import VisuallyHiddenText from 'terra-visually-hidden-text';
 import styles from './Card.module.scss';
 import CardBody from './CardBody';