+++ packages/terra-form-select/tests/jest/MaxSelection.test.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import MaxSelection from '../../src/_MaxSelection';
-import intlContexts from './intl-context-setup';
 
 describe('MaxSelection', () => {
   it('should render a default MaxSelection', () => {
-    const wrapper = shallow(<MaxSelection value="3" />, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(<MaxSelection value="3" />);
     expect(wrapper).toMatchSnapshot();
   });
 });