+++ packages/terra-i18n/src/intl.js
@@ -1,0 +1,4 @@
+import IntlPolyfill from 'intl';
+
+const exportedIntl = typeof (Intl) === 'undefined' ? IntlPolyfill : window.Intl;
+export default exportedIntl;