+++ packages/terra-image/src/terra-dev-site/doc/example/ImageDefault.jsx
@@ -23,9 +23,10 @@ const ImageDefault = () => (
     <br />
     <h2>Placeholder</h2>
     <h3>Invalid src prop (placeholder will show)</h3>
-    <Image src="/mock/failure/route" placeholder={<Image src={placeholderPic1000x200} />} alt="placeholder image" />
+    {/* eslint-disable-next-line react/jsx-no-duplicate-props */}
+    <Image src="/mock/failure/route" alt="invalid src attribute image" placeholder={<Image src={placeholderPic1000x200} alt="placeholder image" />} />
     <h3>Valid src prop (src will show)</h3>
-    <Image src={placeholderPic150x150} placeholder={<Image src={placeholderPic1000x200} />} alt="src image" />
+    <Image src={placeholderPic150x150} alt="src image" placeholder={<Image src={placeholderPic1000x200} alt="placeholder image" />} />
   </div>
 );