+++ packages/terra-list/src/ListSubsection.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import 'terra-base/lib/baseStyles';
 import SubsectionHeader from './ListSubsectionHeader';
 
 const propTypes = {