+++ packages/terra-show-hide/src/_ShowHideButton.jsx
@@ -3,7 +3,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import KeyCode from 'keycode-js';
-import 'terra-base/lib/baseStyles';
 import styles from './_ShowHideButton.module.scss';