+++ packages/terra-table/CHANGELOG.md
@@ -8,6 +8,16 @@ Unreleased
 * Added ARIA label for pre-selected and selected rows
 * Added Translations for selectableTableRows
 
+3.8.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
+3.7.0 - (March 21, 2019)
 ------------------
 ### Changed
 * Replaced local keyCode values with keycode-js npm package